Fix for issue# 8479.
rang and lang changed between HTML 4 and HTML 5. All the rest stayed the
same, hence why those two lines changed. There are plenty of additions
as well, but I left the table as it was (aside from the fix for rang and
lang) and just gave a link to the HTML 5 spec for the full list.