Fix section talking about virtual with regards to package.
It previously indicated that package member functions were virtual,
which is not the case. It also previously indicated that the compiler
knew the entire class hierarchy when compiling the code, which is also
not true (e.g. the base class could be in one library and the derived
class could be in another, and those libraries could be compiled
separately). Both have been fixed.