Remove section in style guide about comments.
We don't really follow the guidelines on comments, and some of them are
arguably outright bad (e.g. using version(none) does not play nicely
with syntax highlighting, since the code appears to be active unlike
when it's commented out). There was some agreement on this when I
brought it up in the past, but for the most part, no one even responded
to attempted discussions on it. But since we don't really follow the
comment guideliness, and they _are_ bad in at least one case, I think
that it's far better to just remove them. The only case where we even
get any dispute about comments at this point is whether ddoc comments
should use /++ +/ or /** */ and the comment guidelines don't discuss
that.