latex.ddoc: Increase premulticols
This fixes the following multicols error:

! Package multicol Error: Error saving partial page.

See the multicol package documentation for explanation.
Type  H <return>  for immediate help.
 ...

l.4182 \begin{multicols}{5}
                               \textit{LinkageAttribute}
? H
The part of the page before the multicols environment was nearly full with
the result that starting the environment will produce an overfull page. Some
text may be lost! Please increase \premulticols either generally or for this
environment by specifying a suitable value in the second optional argument to
the multicols environment.