Adjusted text on compiler support
Provided some additional details and information in getting SIMD types
to work at a compiler level

Fix spelling issue

Fix spelling issue

Fixing multiplication example

In the table below, multiplication is not defined by SIMD int4 types. I
have not been able to get the example to compile as a result. However,
this seems to work.

Removing text related to version(D_SIMD)

Apparently version(D_SIMD) is not supposed to be used in this case. Made some other small modifications to the text to make it flow better with the other stuff removed.

Fixing -m64 flag

Incorporating andralex's recommendation to replace a code line for -m64
for something inline.

Changing Conversion text

Bringing the float4 text back to int4, but changing the multiplication
to addition (because that works).

Removing example for DMD

Having some issues with building this with DMD reference. Just removing
text and keeping it more general.