bugstats.php.dd: Add hidden iframe to trigger generating the image
I've regularly noticed that the graph image linked to from bugstats.php
would not display - however, it would start displaying for a while when
you visit the page it links to. Since the image URL seems to link to a
static file, I guess that the graphs are generated by the CGI request,
and the static files are cleaned up periodically. Thus, to have the image
display on another website (dlang.org), we also need to regularly trigger
its generation, one way being to make the CGI request to generate the
image at the same time as when requesting the image.