add dmd.html
There's a redirect in place that detects the user's operating system and
goes to dmd-{windows,linux,...}.html. The redirect is fine and on the
online dlang.org it completely hides the dmd.html file that's being added
here. The file is needed, because dlang.org is also distributed as a bunch
of .html files as part of the dmd releases.

Not putting links to dmd-windows.html, etc into the file, because they'd
only be forgotten there.