posix.mak: Fix default target
The default target of a Makefile is usually the first defined
non-pattern target. The intended default target of dlang.org's
posix.mak is the 'all' target, however this was broken in
44dd86f9a22bffbc057e76a2cf861037f8f34760 by introducing a non-pattern
target among the pattern targets (in order to specialize building
error pages as necessary to fix relative links to resources).

Move the non-pattern rules block (which includes the intended default
target, 'all') above the pattern rules block to fix the intended
default target.