Change documentation style requirement to /++ +/.
If we're going to require either /++ +/ or /** */, /++ +/ makes more
sense, because it allows for comments to be put into the documentation
if need be, whereas /** */ does not, because it doesn't allow comments
to nest. And while comments would usually be in code examples inside a
documented unit test, sometimes we still need to put code examples
directly in the documentation.