Fix broken twitter embed
Twitter changed their policy and implementation for embedded timelines. They no longer allow embeds of hashtag searches. That broke the twitter box in index.html -- it was showing a link instead. This edit replaces it with an embedded timeline for @D_Programming, which is the best we can do now.