Add LDC to list of snap packages available to download
As presented in my DConf 2019 lightning talk, there is a big discrepancy
in usage between the DMD and DUB snap packages, and LDC.  One might well
assume that this is because the first two are listed on the downoad page
but LDC is not.

This patch corrects the discrepancy, and splits out the install commands
into separate lines with helpful comments.