Revise the Introduction for greater accuracy and reduced redundancy
Currently the introduction has a description of the phases of compilation.
A paragraph describes that each prior phase in unaffected by a subsequent phase.
This description is redundant as this is true for most compilers. Unless
more specific example can be given how D's phases differ from other languages
it is safer to not say anything specific here.

Secondly the phases of compilation goes into unnecessary and random details
of the lexical phase. This is not necessary.

The revised text tries to address these issues. It also adds some information
that may be useful to people who do not know D such as there is no
pre-processing phase.

Signed-off-by: Dibyendu Majumdar <mobile@majumdar.org.uk>