Bring eponymous template spec in line with implementation
According to the spec, a template instantiation only refers to an eponymous member if that member's declaration includes all of the template's parameters. As of DMD v2.090.0, this is not the actual behavior of the implementation, as demonstrated by the following example:

```d
template foo(T)
{
    // Does not include T
    enum foo = 42;
}

// Passes, but should fail according to spec
static assert(foo!void == 42);
```

This change updates the spec to match the actual behavior of the compiler.