[spec/template.dd] Improve TemplateSequenceParameter docs
Any argument passed to an alias parameter can be passed to a sequence
parameter.
Show AliasSeq definition.
Link to `articles/ctarguments`.
`typeof` can be used on a *ValueSeq* to obtain a *TypeSeq*.
Mention a sequence is automatically expanded when referenced in a
declaration or expression.
Split print/write example into 2 - one with ValueSeq arguments, one with
TypeSeq arguments.
A *TypeSeq* can similarly be used to declare lvalue variables.
Add *Sequence Operations* subheading & add example.
Explain how to construct a sequence from an existing one.
Mention AliasAssign for iterative construction.
Mention foreach on a sequence.

type.dd:
`typeof` can be used on a *ValueSeq* to obtain a *TypeSeq*.