Fix 22959 - Extend documentation for C/D main functions.
For D mains, document that parameter / return types may be qualified and
also can be replaced by enums with appropriate base types.

For C mains, document that it may be used without -betterC and how it
interacts with runtime initialization. Replaced the list of allowed
signatures with a proper grammar and also added a workaround for exotic
main signatures.