[spec] Improve is(Type == keyword) docs
Describe `== TypeCtor` more precisely and add example.
Remove is(i == const) example as even if `const int i;` was used it
would still be false because `i` isn't a type.
Add note to prefer isModule and isPackage __traits (ideally `== module`
and `== package` would be deprecated as they only pass when *Type* is
not a type).
For Identifier form, link to non-Identifier to avoid repeating the same
info.