Update function.dd (Lazy Variadic Functions)
Switch the terms "former" and "latter" around. Currently the spec sets "former" to variadic delegate array element access causing evaluation of the other array elements, and "latter" to the lazy variadic array element access only evaluating that element. This is contrary to the example code so we just need to switch the positions of "former" and "latter" in the sentence.