Add macros for keywords. (#4231)
While ddoc will color keywords as keywords in code examples, it won't do
so with $(D ) or ``. To highlight them as keywords, you need to use
$(D_KEYWORD ), but typing out stuff like $(D_KEYWORD typeof) or
$(D_KEYWORD @property) gets pretty verbose, especially if the keyword is
mentioned several times in a paragraph.

So, this adds macros for the keywords so that you don't have to type all
of that out to get the same effect. So, instead of $(D_KEYWORD typeof),
it would be $(K_TYPEOF). Instead of $(D_KEYWORD @property), it would be
$(K_PROPERTY). Etc. Each keyword has a similar macro and is thus
similarly shortened.