Update hash-map.dd
This update is in reference to <https://github.com/dlang/dlang.org/issues/4245>.

The spec currently says that for classes to be used as keys in associative arrays, they *must* override toHash and opEquals. This is untrue because the default implementations use the address of the class instance for computing the hash and for comparisons.