[spec/expression] Improve function literal docs (#4355)
* [spec/expression] Function literal return type is BasicTypeWithSuffixes

* [spec/expression] Improve function literal docs

Move *RefOrAutoRef* into lexical order.
Mention template literal when parameter type omitted, and that a lone
Identifier will be the parameter name.
Move sentence about aliasing literals.
Add examples of literals.
Link to function pointers, delegates (not closures).
Change *Function Literal Aliasing* subheading to *Function Literal
Templates* (change alias link in `declaration.dd`).
A function literal can be a template regardless of aliasing.
The template is not a function template, it's a template of a typed
function literal.
Implicit instantiation is supported when the literal is called.
A literal with `auto ref` parameter cannot be explicitly instantiated.